+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/about/AndroidAboutView.kt
@@ -2,14 +2,22 @@
 
 package co.electriccoin.zcash.ui.screen.about
 
+import android.content.Context
 import androidx.activity.ComponentActivity
+import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import co.electriccoin.zcash.configuration.AndroidConfigurationFactory
 import co.electriccoin.zcash.ui.MainActivity
+import co.electriccoin.zcash.ui.R
 import co.electriccoin.zcash.ui.common.model.VersionInfo
+import co.electriccoin.zcash.ui.screen.about.util.WebBrowserUtil
 import co.electriccoin.zcash.ui.screen.about.view.About
 import co.electriccoin.zcash.ui.screen.support.model.ConfigInfo
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.launch
 
 @Composable
 internal fun MainActivity.WrapAbout(goBack: () -> Unit) {
@@ -29,9 +37,37 @@ internal fun WrapAbout(
         AndroidConfigurationFactory.getInstance(activity.applicationContext).hintToRefresh()
     }
 
+    val snackbarHostState = remember { SnackbarHostState() }
+    val scope = rememberCoroutineScope()
+
     About(
         onBack = goBack,
         versionInfo = versionInfo,
-        configInfo = configInfo
+        configInfo = configInfo,
+        onPrivacyPolicy = {
+            openPrivacyPolicyInWebBrowser(
+                activity.applicationContext,
+                snackbarHostState,
+                scope
+            )
+        },
+        snackbarHostState = snackbarHostState,
     )
+}
+
+fun openPrivacyPolicyInWebBrowser(
+    context: Context,
+    snackbarHostState: SnackbarHostState,
+    scope: CoroutineScope
+) {
+    val storeIntent = WebBrowserUtil.newActivityIntent(WebBrowserUtil.ZCASH_PRIVACY_POLICY_URI)
+    runCatching {
+        context.startActivity(storeIntent)
+    }.onFailure {
+        scope.launch {
+            snackbarHostState.showSnackbar(
+                message = context.getString(R.string.about_unable_to_web_browser)
+            )
+        }
+    }
 }