+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/securitywarning/view/SecurityScreenTag.kt
@@ -5,5 +5,4 @@ package co.electriccoin.zcash.ui.screen.securitywarning.view
  */
 object SecurityScreenTag {
     const val ACKNOWLEDGE_CHECKBOX_TAG = "acknowledge_checkbox"
-    const val WARNING_TEXT_TAG = "warning_text"
 }